+++ inferno-ml-server/exe/ParseAndSave.hs
@@ -41,7 +41,8 @@ import Inferno.Core
   ( Interpreter (Interpreter, parseAndInfer),
     mkInferno,
   )
-import Inferno.ML.Server.Module.Prelude (mkBridgePrelude)
+import Inferno.ML.Module.Prelude (mlPrelude)
+import Inferno.ML.Server.Module.Prelude (mkServerBridgePrelude)
 import Inferno.ML.Types.Value (customTypes)
 import Inferno.Types.Syntax (Expr, TCScheme)
 import Inferno.Types.VersionControl
@@ -81,7 +82,7 @@ parseAndSave ipid p conns ios = do
   now <- fromIntegral @Int . round <$> getPOSIXTime
   ast <-
     either (throwString . displayException) pure . (`parse` t)
-      =<< mkInferno @_ @BridgeMlValue (mkBridgePrelude funs) customTypes
+      =<< mkInferno @_ @BridgeMlValue (mkServerBridgePrelude funs mlPrelude) customTypes
   bracket (connectPostgreSQL conns) close (saveScriptAndParam ipid ast now ios)
 
 saveScriptAndParam ::