+++ inferno-ml-server/src/Inferno/ML/Server/Module/Prelude.hs
@@ -5,12 +5,14 @@
 
 module Inferno.ML.Server.Module.Prelude
   ( bridgeModules,
-    mkBridgePrelude,
+    mkServerBridgePrelude,
+    serverMlPrelude,
   )
 where
 
+import Control.Exception (ErrorCall)
 import Control.Monad.Catch (MonadCatch, MonadThrow (throwM))
-import Control.Monad.IO.Class (MonadIO)
+import Control.Monad.IO.Class (MonadIO (liftIO))
 import Data.Foldable (foldrM)
 import Data.Int (Int64)
 import qualified Data.IntMap as IntMap
@@ -18,20 +20,29 @@ import qualified Data.Map as Map
 import Data.Tuple.Extra ((&&&))
 import Foreign.C (CTime (CTime))
 import Inferno.Eval.Error (EvalError (RuntimeError))
-import Inferno.ML.Module.Prelude (mlPrelude)
+import Inferno.ML.Module.Prelude (getDevice, mkMlPrelude)
 import Inferno.ML.Server.Module.Types
-import Inferno.ML.Server.Types (IValue)
+import Inferno.ML.Server.Types
+  ( IValue,
+    RemoteM,
+    TraceWarn (CouldntMoveTensor),
+    logWarn,
+  )
 import Inferno.ML.Types.Value (MlValue (VExtended), mlQuoter)
 import Inferno.Module.Cast
 import Inferno.Module.Prelude (ModuleMap)
 import Inferno.Types.Syntax (ExtIdent (ExtIdent))
 import Inferno.Types.Value
   ( ImplEnvM,
-    Value (VArray, VCustom, VEmpty, VEpochTime, VFun, VTuple),
+    Value (VArray, VCustom, VEmpty, VEnum, VEpochTime, VFun, VTuple),
+    liftImplEnvM,
   )
 import Inferno.Types.VersionControl (VCObjectHash)
 import Lens.Micro.Platform
 import System.Posix.Types (EpochTime)
+import Torch (Device, Tensor)
+import qualified Torch (toDevice)
+import UnliftIO.Exception (handle)
 
 -- | Contains primitives for use in bridge prelude, including those to read\/write
 -- data
@@ -205,39 +216,41 @@ module DataSource
             VTuple [VEpochTime t, x] -> (t,) <$> toIValue x
             _ -> throwM $ RuntimeError "extractPair: expected a tuple (time, 'a)"
 
-mkBridgePrelude ::
+-- | Make a prelude that works without @RemoteM@ monad
+mkServerBridgePrelude ::
   forall m.
   ( MonadIO m
-  , MonadThrow m
   , MonadCatch m
+  , MonadThrow m
   ) =>
   BridgeFuns m ->
+  ModuleMap m BridgeMlValue ->
   ModuleMap m BridgeMlValue
-mkBridgePrelude bfuns =
+mkServerBridgePrelude bfuns mlPrelude =
   case modules & view (at "Base") &&& view (at "DataSource") of
     (Just base, Just source) ->
       modules
         & at "DataSource"
-        .~ Nothing
+          .~ Nothing
         & at "Base"
-        ?~ ( base
-              & #moduleOpsTable
-              %~ flip
-                (IntMap.unionWith (<>))
-                (view #moduleOpsTable source)
-              & #moduleTypeClasses
-              <>~ view #moduleTypeClasses source
-              & #moduleObjects
-              . _1
-              <>~ view (#moduleObjects . _1) source
-              & #moduleObjects
-              . _2
-              <>~ view (#moduleObjects . _2) source
-              & #moduleObjects
-              . _3
-              %~ (`combineTermEnv` view (#moduleObjects . _3) source)
-           )
-    _ -> error "mkBridgePrelude: Missing Base and/or DataSource modules"
+          ?~ ( base
+                & #moduleOpsTable
+                  %~ flip
+                    (IntMap.unionWith (<>))
+                    (view #moduleOpsTable source)
+                & #moduleTypeClasses
+                  <>~ view #moduleTypeClasses source
+                & #moduleObjects
+                  . _1
+                  <>~ view (#moduleObjects . _1) source
+                & #moduleObjects
+                  . _2
+                  <>~ view (#moduleObjects . _2) source
+                & #moduleObjects
+                  . _3
+                  %~ (`combineTermEnv` view (#moduleObjects . _3) source)
+             )
+    _ -> error "mkServerBridgePrelude: Missing Base and/or DataSource modules"
   where
     combineTermEnv ::
       (Map.Map ExtIdent v, Map.Map VCObjectHash e) ->
@@ -247,7 +260,56 @@ mkBridgePrelude bfuns =
 
     modules :: ModuleMap m BridgeMlValue
     modules =
-      Map.unionWith
-        (error "Redefined builtins")
-        (mlPrelude @_ @BridgeValue)
-        $ bridgeModules @_ bfuns
+      Map.unionWith (error "Redefined builtins") mlPrelude $
+        bridgeModules @_ bfuns
+
+-- | ML prelude for use only in @RemoteM@ (needed for tracing effects)
+serverMlPrelude :: ModuleMap RemoteM (MlValue BridgeValue)
+serverMlPrelude = mkMlPrelude toDeviceFun
+  where
+    toDeviceFun ::
+      Value (MlValue BridgeValue) (ImplEnvM RemoteM (MlValue BridgeValue))
+    toDeviceFun =
+      VFun $ \case
+        VEnum _ e ->
+          -- The device that the user wants to move the tensor to; currently
+          -- can only be `#cpu` or `#cuda`
+          getDevice e <&> \device ->
+            -- Original tensor to be moved (as Inferno value)
+            VFun $ \vtensor -> do
+              fromValue vtensor >>= liftIO . toDeviceIO device >>= \case
+                -- This is the (potentially) moved tensor; its having been moved
+                -- or not depends on the devices involved
+                --
+                -- In most cases, this would happen if a CPU-only `inferno-ml-server`
+                -- tries to evaluate `ML.toDevice t #cuda`; with no CUDA device
+                -- it can't be moved. Logging is more important, however, if
+                -- for some reason CUDA-enabled `inferno-ml-server`s can't move
+                -- the device when they should be able to. Logging can at least
+                -- help with detecting these types of problems
+                --
+                -- It's better to _not_ throw an exception in the `Left` case
+                -- here so the same scripts can be used without modification
+                -- on both CPU and GPU `inferno-ml-server`s
+                Right tensor -> pure $ toValue @_ @_ @Tensor tensor
+                Left tensor -> do
+                  liftImplEnvM . logWarn . CouldntMoveTensor $ show device
+                  pure $ toValue @_ @_ @Tensor tensor
+        _ -> throwM $ RuntimeError "toDeviceFun: expecting a device enum"
+
+-- Workaround for `error`s in Torch's "pure" `toDevice`. If the original
+-- tensor cannot be moved, `Left Tensor` is returned to signal failure to move
+-- the tensor to the new device, i.e. the original tensor is returned;
+-- `Right Tensor` is the successfully moved new tensor
+toDeviceIO :: Device -> Tensor -> IO (Either Tensor Tensor)
+toDeviceIO device t1 = handle handler $ t2 `seq` pure (Right t2)
+  where
+    -- If the `error` occurs, the original tensor is returned
+    handler :: ErrorCall -> IO (Either Tensor Tensor)
+    handler = const . pure $ Left t1
+
+    -- The potentialy moved tensor. If the requested device is not available,
+    -- i.e. moving to CUDA on a CPU-only device, this "pure" function will
+    -- throw an `error`
+    t2 :: Tensor
+    t2 = Torch.toDevice device t1